(GH-1867) Document how privilege escalation works in Bolt
This adds a net new page to the docs titled `Escalating privilege with
Bolt` which details how Bolt builds the command to change users, how to
configure that command, and what limitations Bolt has with respect to
escalating privilege.

The page also includes a list of escalation-related configuration
options and command line flags, which are generated from Bolt itself the
same way the command line reference and configuration reference pages
are generated. The generation of the options on that page is done as
part of the `docs:config_reference` rake task. In order for this to
work, the run-as configuration options stored in
`Bolt::Config::Transport::SSH::OPTIONS` and
`Bolt::Config::Transport::Local::OPTIONS` had to be moved from those
constants to `Bolt::Config::Transport::Base::RUN_AS_OPTIONS` to be read
in the Rakefile. Those options are now merged into the other transport
options defined in those constants when the constant is instantiated.

!no-release-note

Closes #1867